requirements.txt
asgiref==3.7.2
distlib==0.3.5
Django==5.0.1
django-multiselectfield==0.1.12
django-phonenumber-field==7.3.0
filelock==3.7.1
phonenumbers==8.13.27
pillow==10.2.0
platformdirs==2.5.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
virtualenv==20.16.2
